wip
Former-commit-id: 63664dfbf2c166ec9a2fa3ec314f85de8142049a
Former-commit-id: 673e4844d1aa838bb07117a8bbe8407de1612dc5

+++ framework/StandAloneUtilityClasses/XMLArchive/Headers/XLSXDocument.h
@@ -54,7 +54,7 @@ protected:
 		if (f == mStringIndex.end())
 		{
 			mSharedStrings.push_back(s);
-			result = mSharedStrings.size() - 1;
+			result = (u32)mSharedStrings.size() - 1;
 			mStringIndex[s] = result;
 		}
 		else
